feature_pipeline/pyproject.toml
[tool.poetry]
name = "feature-pipeline"
version = "0.1.0"
description = ""
authors = ["Plouplou37 <quentin.plourdeau@gmail.com>"]
readme = "README.md"
packages = [{include = "feature_pipeline"}]

[tool.poetry.dependencies]
python = "~3.9"
hopsworks = "3.4.3"
fire = "^0.5.0"
yarl = "^1.8.2"
pandas = ">=1.3.5"
requests = "^2.28.2"
python-dotenv = ">=0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
